+++ main.go
@@ -1,15 +1,7 @@
 package main
 
-import (
-	"os"
-
-	"github.com/vulcanshen/clerk/cmd"
-	"github.com/vulcanshen/clerk/internal/platform"
-)
+import "github.com/vulcanshen/clerk/cmd"
 
 func main() {
-	if len(os.Args) > 1 && (os.Args[1] == "feed" || os.Args[1] == "punch") {
-		platform.HideConsoleWindow()
-	}
 	cmd.Execute()
 }